FIX: fixed tests after merging
